Fix: Allow sidebars to scroll independently of main window
When the length of the sidebar exceeded that of the viewport, the body
would scroll. This lead to unusual scroll behaviour in the console as
both the console container and the main body would scroll at the same
time.

Resolve by fixing the height of the sidebars to 100% and allowing them
to scroll if necessary. This means it is never necessary for the body to
scroll and so the two scrolling areas now feel independent of each
other.

Closes https://github.com/Snapmaker/Luban/issues/500